Changed the sampling method
This commit is about changing the sampling method. Instead of using the function 'sample.split' from the caTools library, the function 'sample' is used from the core R library. This is done because it removes the dependency on a external library which has to be installed and loaded and to make the sampling more insightful.
- line 18 and 53: removal of caTools library installation and loading
- line 68: define number of molecules that will be preserved for the test set
- line 73: using the sample, seq_len and nrow functions the set of rownumbers that will be in the test set is defined
- line 75 and 76: here the actual data is divided into a test and training set (they have no overlap between each other)
- line 86: making sure that the right boiling point column from the training set is used as the expected y (this was unnoticed before)