Documentation, variable names and references
This commit's main component is improved documentation. For all the sections of code a short explanation of what that section is about is included to give the user the most direct link to the code.
- line 81: A new section is made of the part that contains the splitting of the data in a training and a test set. Before this was included in the smiles section. This step is taken to improve the modularity of the code and to make sure to the user that the methods used in both sections have a different background.

Variables:
line 32: The variable name of the dataset is changed from  'r' to 'alkanes' as this is a more descriptive name for the dataset.
line 74: Also the variable 'data_set' is changed to 'bp_and_descs' for the same reason.

References:
line 112: A reference section is included containing papers where the used descriptors are defined. This is done to give the user a link to the background of the descriptors.