User explaination for running the code
This commit includes the necessary explanation for the user to make sure that the user has the right version of Java installed, so that the rJava package can be loaded. (line 57)
Extra explanation that the user is aware that the random seed is set at 5, this is important if the user wants to reproduce the same results. (line 83)