Introduction and user manual
The introduction to the project and the user manual are described. This is to make the user aware of the goal of this project and the how she/he has to treat the code.

The author description is included in this file for now. As there is only one author it is reasonable to include the author list here.

References that are used in the code and a link to the course description are given in the references.

Furthermore, a installation section is preserved for installation instructions, these will be included later.