Null model error
The error of the null model is calculated (line131) and plotted (line 159-163). Therefore the package Metrics is needed and is included in the installation requirements and in the references.
Furthermore variable names are improved of the error objects, to make the it better linked to the content.
Also the label of the first plot is changed to make it consistent with the content of the figure.